refactor(web): update user imports to use types instead of schema
