+++ web/src/routes/global/auth.data.ts
@@ -1,4 +1,4 @@
-import type { User } from "../../schema/user";
+import type { User } from "../../types/user";
 import { makeLoader } from "react-router-typesafe";
 
 export const authLoader = makeLoader(async () => {
