+++ web/src/types/search.ts
